small refactor on logging
