+++ src/account/TransferTransaction.js
@@ -411,6 +411,7 @@ export default class TransferTransaction extends Transaction {
      * @returns {TransferTransaction}
      */
     addHbarTransfer(accountId, amount) {
+        this._logger.debug("adding hbar transfer...");
         return this._addHbarTransfer(accountId, amount, false);
     }
 
@@ -702,6 +703,7 @@ export default class TransferTransaction extends Transaction {
      * @returns {Promise<HashgraphProto.proto.ITransactionResponse>}
      */
     _execute(channel, request) {
+        this._logger.debug("executing transfer transaction")
         return channel.crypto.cryptoTransfer(request);
     }
