+++ src/topic/TopicCreateTransaction.js
@@ -182,7 +182,7 @@ export default class TopicCreateTransaction extends Transaction {
     setTopicMemo(topicMemo) {
         this._requireNotFrozen();
         this._topicMemo = topicMemo;
-
+        this._logger.debug("setting topic memo...");
         return this;
     }
